+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/datastructures/networking/libp2p/rpc/BeaconBlocksByRootRequestMessageTest.java
@@ -48,7 +48,7 @@ class BeaconBlocksByRootRequestMessageTest {
   @Test
   public void verifyMaxLengthOfContainerIsGreaterOrEqualToMaxRequestBlocks() {
     assertThat(schema.getMaxLength())
-        .isGreaterThanOrEqualTo(spec.getNetworkingConfig().getMaxRequestBlocks().longValue())
+        .isGreaterThanOrEqualTo(spec.getNetworkingConfig().getMaxRequestBlocks())
         .isGreaterThanOrEqualTo(MAX_REQUEST_BLOCKS_DENEB.longValue());
   }
 }
